+++ src/webgpu/index.js
@@ -1,4 +1,7 @@
 export * from './bvh_ray_functions.wgsl.js';
 export * from './common_functions.wgsl.js';
 export * from './distance_functions.wgsl.js';
-export { BVHComputeData } from './BVHComputeData.js';
+export * from './BVHComputeData.js';
+export * from './tsl/structs.js';
+export * from './tsl/fns.js';
+export * from './tsl/constants.js';
